requirements.txt
websocket-client>=1.5.2
python-dotenv>=1.0.0
